change dob to age
