+++ src/main/java/seedu/address/model/applicant/Age.java
@@ -1,0 +1,48 @@
+package seedu.address.model.applicant;
+
+import static java.util.Objects.requireNonNull;
+import static seedu.address.commons.util.AppUtil.checkArgument;
+
+public class Age {
+    public static final String MESSAGE_CONSTRAINTS =
+            "Age should only contain numbers, and it should be at least 2 digits long";
+    public static final String VALIDATION_REGEX = "\\d{2,}";
+    public final String age;
+
+    /**
+     * Constructs an {@code age}.
+     *
+     * @param age A valid age .
+     */
+    public Age(String age) {
+        requireNonNull(age);
+        checkArgument(isValidAge(age), MESSAGE_CONSTRAINTS);
+        this.age = age;
+    }
+
+    /**
+     * Returns true if a given string is a valid age.
+     */
+    public static boolean isValidAge(String test) {
+        return test.matches(VALIDATION_REGEX);
+    }
+
+    @Override
+    public String toString() {
+        return age;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof Gender // instanceof handles nulls
+                && age.equals(((Age) other).age)); // state check
+    }
+
+    @Override
+    public int hashCode() {
+        return age.hashCode();
+    }
+
+
+}
