+++ src/main/java/seedu/address/model/applicant/Gender.java
@@ -9,8 +9,8 @@ import static seedu.address.commons.util.AppUtil.checkArgument;
  */
 public class Gender {
 
-    public static final String MESSAGE_CONSTRAINTS =  "Gender should be exactly one character and only M or F " +
-                                                        "(case sensitive)";
+    public static final String MESSAGE_CONSTRAINTS =
+            "Gender should be exactly one character and only M or F (case sensitive)";
     public static final String VALIDATION_REGEX = "M|F";
     public final String value;
